refactor: convertLegacyProps function moved to buttonHelpers
