+++ components/_util/ActionButton.tsx
@@ -2,7 +2,7 @@ import useState from 'rc-util/lib/hooks/useState';
 import * as React from 'react';
 import Button from '../button';
 import type { ButtonProps, LegacyButtonType } from '../button/button';
-import { convertLegacyProps } from '../button/button';
+import { convertLegacyProps } from '../button/buttonHelpers';
 
 export interface ActionButtonProps {
   type?: LegacyButtonType;
