+++ components/button/buttonHelpers.tsx
@@ -1,8 +1,18 @@
 import React from 'react';
 import { cloneElement, isFragment } from '../_util/reactNode';
+import type { BaseButtonProps, LegacyButtonType } from './button';
 
 const rxTwoCNChar = /^[\u4e00-\u9fa5]{2}$/;
 export const isTwoCNChar = rxTwoCNChar.test.bind(rxTwoCNChar);
+
+export function convertLegacyProps(
+  type?: LegacyButtonType,
+): Pick<BaseButtonProps, 'danger' | 'type'> {
+  if (type === 'danger') {
+    return { danger: true };
+  }
+  return { type };
+}
 
 export function isString(str: any): str is string {
   return typeof str === 'string';
