+++ components/modal/components/NormalOkBtn.tsx
@@ -2,7 +2,7 @@ import type { FC } from 'react';
 import React, { useContext } from 'react';
 
 import Button from '../../button';
-import { convertLegacyProps } from '../../button/button';
+import { convertLegacyProps } from '../../button/buttonHelpers';
 import { ModalContext } from '../context';
 import type { ModalProps } from '../interface';
